refactor 1.0
